# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "La documentacion de mi ecommerce",
    "version": "1.0.0",
    "title": "Videogame Ecommerce",
    "termsOfService": "http://videogameecommerce.herokuapp.com/terms/",
    "contact": {
      "email": "lalo0892@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "videogameecommerce.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/products": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Add new product",
        "description": "Agrega un nuevo producto a la base de datos de ecommerce",
        "operationId": "addProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "El producto que hay que agregar",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get all products",
        "description": "",
        "operationId": "getAllProducts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "platform": {
          "type": "string",
          "enum": [
            "Xbox",
            "PS4",
            "Nintendo"
          ]
        },
        "description": {
          "type": "string"
        }
      }
    },
...
  }
}